Update to move vitest workspace into existing vite config.
